+++ src/ui/ClamPopup.hpp
@@ -20,6 +20,10 @@ protected:
     void rebuildNearbyGames();
     void appendConsole(std::string const& line);
 
+    float m_scrollWidth = 280.f;
+    float m_nearbyScrollHeight = 72.f;
+    float m_consoleScrollHeight = 32.f;
+
     cocos2d::CCLabelBMFont* m_statusLabel = nullptr;
     cocos2d::CCLabelBMFont* m_peerLabel = nullptr;
     cocos2d::CCLabelBMFont* m_consoleLabel = nullptr;
@@ -27,7 +31,7 @@ protected:
 
     std::string m_consoleText;
     std::string m_nearbyFingerprint;
-    size_t m_maxConsoleLines = 10;
+    size_t m_maxConsoleLines = 6;
 
 public:
     static ClamPopup* create();